requirements.txt
streamlit>=1.29.0
streamlit-mermaid>=0.1.1
openai>=1.3.5
Pillow>=10.1.0
pdf2image>=1.16.3
PyMuPDF>=1.23.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
typing_extensions>=4.8.0
numpy>=1.24.3
boto3>=1.34.0